Update test coverage for the Caregivers app submit transformer
